// dashboard/src/features/cashier-management-dashboard/interfaces/Branch.tsx
export interface Branch {
  branchAddress: string;
  branchContact: string;
  branchCreatedBy: string;
  branchCreatedOn: string;
  branchDescription: string;
  branchEmail: string;
  branchFax: string;
  branchId: number;
  branchImage: string | null;
  branchLocation: string;
  branchName: string;
  branchStatus: boolean;
}